+++ src/api/interfaces/CRUDApi.ts
@@ -2,6 +2,8 @@ export default interface CRUDApi {
 
     fetchAll(orderBy: String, orderWay: String, page: Number, perPage: Number): Promise<String>;
 
+    fetchSingle(id: String): Promise<Object>;
+
     post(data: Object): Promise<Object|String>;
     
 }