+++ src/js/reducers/app/roles.js
@@ -6,7 +6,8 @@ const initialState = {
 
 const roles = (state = initialState, action) => {
     switch (action.type) {
-        case "TRY_FETCH_ROLES_COLLECTION": {
+        case "TRY_FETCH_ROLES_COLLECTION":
+        case "TRY_FETCH_ROLES": {
             return Object.assign({}, state, {
                 isLoading: true
             });
@@ -20,7 +21,15 @@ const roles = (state = initialState, action) => {
             });
         }
 
-        case "FETCH_ROLES_COLLECTION_FAILED": {
+        case "FETCH_ROLES_SUCCEEDED": {
+            return Object.assign({}, state, {
+                isLoading: false,
+                role: action.data ? action.data : null
+            });
+        }
+
+        case "FETCH_ROLES_COLLECTION_FAILED":
+        case "FETCH_ROLES_FAILED": {
             return Object.assign({}, state, {
                 isLoading: false,
                 error: action.error