+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Next version
 + Add refresh token to Dashboard
++ Add single fetch. Prepare for call detail
 
 ## 0.16.0
 + Add dashboard